refactor: Remove companySchemas and update imports to use openapi/schemas
